# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Air Smart IoT App API Service",
        "title": "Air IoT API Service 2023",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/devices": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get device list",
                "tags": [
                    "ThingsDevice"
                ],
                "summary": "Get Device list",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.ApiResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ApiResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "user is a new things device",
                "tags": [
                    "ThingsDevice"
                ],
                "summary": "New a Things Device",
                "parameters": [
                    {
                        "description": "a device information",
                        "name": "DeviceInfo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/services.Device"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.ApiResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ApiResponse"
                        }
                    }
                }
            }
        },
        "/devices/:id": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update my things device",
                "tags": [
                    "ThingsDevice"
                ],
                "summary": "Update my things device",
                "parameters": [
                    {
                        "description": "Update things device",
                        "name": "DeviceInfo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/services.ReqUpdateDevice"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.ApiResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ApiResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a things device",
                "tags": [
                    "ThingsDevice"
                ],
                "summary": "Delete a things device",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Device id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.ApiResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ApiResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "services.Device": {
            "type": "object",
            "required": [
                "model",
                "name",
                "serialNo",
                "title",
                "warranty"
            ],
            "properties": {
                "model": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "serialNo": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                },
                "warranty": {
                    "type": "string"
                }
            }
        },
        "services.ReqUpdateDevice": {
            "type": "object",
            "required": [
                "model",
                "name",
                "serialNo",
                "title",
                "warranty"
            ],
            "properties": {
                "model": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "serialNo": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                },
                "warranty": {
                    "type": "string"
                }
            }
        },
...
        "utils.ApiResponse": {
            "type": "object",
            "properties": {
                "message": {},
                "status": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}